ml-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
numpy==2.1.3
opencv-python==4.10.0.84
Pillow==11.0.0
torch==2.4.1
torchvision==0.19.1
transformers==4.46.2
huggingface-hub==0.26.2
python-multipart==0.0.17
python-dotenv==1.0.1
onnxruntime-gpu==1.19.2

